breaking: add support for multiple targets (#10)
When building for multiple targets, sbom and vex files are overwritten
by each target and you end up with only the last one.
Fixing by using a dedicated folder for each target.

This is a breaking change as the resulting sbom and vex are one folder
down in the hierarchy.

---------

Co-authored-by: Jasper Ben Orschulko <jasper@fancydomain.eu>