Include component licence info
Add component license information (as stated in the yocto recipe)
into the generated SBOM.

If the recipe license contains "&" or "|", we assume this is a expression,
adding it to licenses.expression.
If the recipe license matches a singular known SPDX license identifier, the
license ID is added to licenses[0].license.id. Otherwise, it is added as
licenses[0].license.name

Signed-off-by: Jasper Orschulko <jasper@fancydomain.eu>